refactor(vue): Change mixins structure.
